Move resolve_versions into new file
The new core/dependency file will contain pubgrub
logic ported from hexpm_rust
