+++ compiler-core/src/lib.rs
@@ -57,6 +57,7 @@ pub mod bit_string;
 pub mod build;
 pub mod codegen;
 pub mod config;
+pub mod dependency;
 pub mod diagnostic;
 pub mod docs;
 pub mod erlang;
